feat: Types supplies support enable/disable

+++ src/main/java/com/ai/st/microservice/providers/business/ProviderBusiness.java
@@ -7,7 +7,6 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.ai.st.microservice.providers.dto.ExtensionDto;
 import com.ai.st.microservice.providers.dto.ProviderAdministratorDto;
 import com.ai.st.microservice.providers.dto.ProviderCategoryDto;
 import com.ai.st.microservice.providers.dto.ProviderDto;
@@ -176,7 +175,8 @@ public class ProviderBusiness {
 		return providerDto;
 	}
 
-	public List<TypeSupplyDto> getTypesSuppliesByProviderId(Long providerId) throws BusinessException {
+	public List<TypeSupplyDto> getTypesSuppliesByProviderId(Long providerId, Boolean onlyActive)
+			throws BusinessException {
 
 		List<TypeSupplyDto> listTypeSupplyDtos = new ArrayList<TypeSupplyDto>();
 
@@ -189,31 +189,11 @@ public class ProviderBusiness {
 		List<TypeSupplyEntity> listTypeSupplyEntities = providerEntity.getTypesSupplies();
 		for (TypeSupplyEntity typeSupplyEntity : listTypeSupplyEntities) {
 
-			TypeSupplyDto typeSupplyDto = new TypeSupplyDto();
-			typeSupplyDto.setCreatedAt(typeSupplyEntity.getCreatedAt());
-			typeSupplyDto.setDescription(typeSupplyEntity.getDescription());
-			typeSupplyDto.setId(typeSupplyEntity.getId());
-			typeSupplyDto.setMetadataRequired(typeSupplyEntity.getIsMetadataRequired());
-			typeSupplyDto.setModelRequired(typeSupplyEntity.getIsModelRequired());
-			typeSupplyDto.setName(typeSupplyEntity.getName());
-
-			List<ExtensionEntity> extensions = typeSupplyEntity.getExtensions();
-			List<ExtensionDto> extensionsDto = new ArrayList<ExtensionDto>();
-			for (ExtensionEntity item : extensions) {
-				ExtensionDto extensionDto = new ExtensionDto();
-				extensionDto.setId(item.getId());
-				extensionDto.setName(item.getName());
-				extensionsDto.add(extensionDto);
+			if ((onlyActive && typeSupplyEntity.getActive()) || !onlyActive) {
+				TypeSupplyDto typeSupplyDto = typeSupplyBusiness.entityParseDto(typeSupplyEntity);
+				listTypeSupplyDtos.add(typeSupplyDto);
 			}
-			typeSupplyDto.setExtensions(extensionsDto);
 
-			ProviderProfileDto providerProfileDto = new ProviderProfileDto();
-			providerProfileDto.setDescription(typeSupplyEntity.getProviderProfile().getDescription());
-			providerProfileDto.setId(typeSupplyEntity.getProviderProfile().getId());
-			providerProfileDto.setName(typeSupplyEntity.getProviderProfile().getName());
-			typeSupplyDto.setProviderProfile(providerProfileDto);
-
-			listTypeSupplyDtos.add(typeSupplyDto);
 		}
 
 		return listTypeSupplyDtos;
@@ -818,6 +798,18 @@ public class ProviderBusiness {
 		ProviderEntity providerEntity = providerService.getProviderById(providerId);
 		if (!(providerEntity instanceof ProviderEntity)) {
 			throw new BusinessException("El proveedor no existe.");
+		}
+
+		if (providerEntity.getTypesSupplies().size() > 0) {
+			throw new BusinessException("No se puede eliminar el proveedor porque ya tiene configurado insumos.");
+		}
+
+		if (providerEntity.getProfiles().size() > 0) {
+			throw new BusinessException("No se puede eliminar el proveedor porque ya tiene áreas registradas.");
+		}
+
+		if (providerEntity.getAdministrators().size() > 0 || providerEntity.getUsers().size() > 0) {
+			throw new BusinessException("No se puede eliminar el proveedor porque ya tiene usuarios registrados.");
 		}
 
 		providerService.deleteProvider(providerEntity);
